+++ src/prefect/assets/__init__.py
@@ -1,0 +1,4 @@
+from prefect.assets.core import Asset
+from prefect.assets.materialize import materialize
+
+__all__ = ["Asset", "materialize"]
