+++ src/prefect/assets/materialize.py
@@ -1,0 +1,38 @@
+from __future__ import annotations
+
+from typing import Any, Callable, TypeVar, Union
+
+from typing_extensions import ParamSpec
+
+from prefect.assets.core import Asset
+from prefect.tasks import Task
+
+T = TypeVar("T")
+P = ParamSpec("P")
+R = TypeVar("R")
+
+
+def materialize(
+    *assets: Union[str, Asset],
+    **task_kwargs: Any,
+) -> Callable[[Callable[P, R]], Task[P, R]]:
+    """
+    Decorator for materializing assets.
+
+    Args:
+        *assets: Assets to materialize
+        **task_kwargs: Additional task configuration
+    """
+    if not assets:
+        raise TypeError(
+            "materialize requires at least one asset argument, e.g. `@materialize(asset1)`"
+        )
+
+    asset_objs = [Asset(key=a) if isinstance(a, str) else a for a in assets]
+
+    def decorator(fn: Callable[P, R]) -> Task[P, R]:
+        task = Task(fn=fn, **task_kwargs)
+        task.assets = asset_objs
+        return task
+
+    return decorator
